+++ spotify/artist.go
@@ -1,0 +1,30 @@
+package spotify
+
+import "fmt"
+
+// Artist represents a spotify Artist
+type Artist struct {
+	Genres     []string `json:"genres"`
+	Name       string   `json:"name"`
+	Popularity int32    `json:"popularity"`
+	URI        string   `json:"uri"`
+	ID         string   `json:"ID"`
+}
+
+// Artists is a collection of spotify Artist
+type Artists []Artist
+
+// EmbeddedPlayer  will return the html to use for rendering the embedded spotify
+// player iframe
+func (a *Artist) EmbeddedPlayer() string {
+	return fmt.Sprintf(`<h4 width="300" style="text-align:center">%s</h4><iframe src="https://open.spotify.com/embed/artist/%s" width="300" height="380" frameborder="0" allowtransparency="true" allow="encrypted-media"></iframe>`, a.Name, a.ID)
+}
+
+// IDs returns the ID for each of the artists in the collection of Artists
+func (a *Artists) IDs() []string {
+	ret := []string{}
+	for _, i := range *a {
+		ret = append(ret, i.ID)
+	}
+	return ret
+}
