+++ spotify/tokenResponse.go
@@ -1,0 +1,19 @@
+package spotify
+
+import "fmt"
+
+type tokenResponse struct {
+	AccessToken  string `json:"access_token"`
+	Type         string `json:"token_type"`
+	Scope        string `json:"scope"`
+	Exp          int    `json:"expires_in"`
+	RefreshToken string `json:"refresh_token"`
+}
+
+func (sr *tokenResponse) ToString() string {
+	ret := ""
+	ret += "-- Spotify Response\n"
+	ret += fmt.Sprint("\tAccessToken: ", sr.AccessToken, "\n")
+	ret += fmt.Sprint("\tScope: ", sr.Scope, "\n")
+	return ret
+}
